requirements.txt
antlr4-python3-runtime==4.8
astroid==2.4.2
certifi==2021.5.30
hydra-core==1.1.0
importlib-resources==5.2.0
isort==5.6.4
lazy-object-proxy==1.4.3
numpy==1.21.0
omegaconf==2.1.0
pip==21.1.3
pylint==2.6.0
PyYAML==5.4.1
setuptools==52.0.0.post20210125
toml==0.10.2
torch==1.9.0
typing-extensions==3.10.0.0
wheel==0.36.2
wrapt==1.12.1
zipp==3.5.0
